+++ .expo/settings.json
@@ -3,7 +3,7 @@
   "lanType": "ip",
   "dev": true,
   "minify": false,
-  "urlRandomness": null,
+  "urlRandomness": "4d-php",
   "https": false,
   "scheme": null,
   "devClient": false
