+++ components/CastField.jsx
@@ -2,29 +2,33 @@ import React, { useState } from 'react';
 import { View, Picker, Switch, Text, TextInput, Alert, ScrollView, Keyboard ,StyleSheet, SafeAreaView} from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { TouchableOpacity } from 'react-native-gesture-handler';
+
 import { createEvent } from '../api/firebaseMethods';
+import { editProfile, uploadPicture } from '../api/firebaseMethods';
+
 import * as ImagePicker from 'expo-image-picker';
 import * as ImageManipulator from 'expo-image-manipulator';
-import { editProfile, uploadPicture } from '../api/firebaseMethods';
+
 import DefaultImg from '../assets/images/defaultprofile.png'
 
-export default Cast = ({cast, castName, castRole}) => {
+export default Cast = ({cast, index, onTextUpdate}) => {
 
   return (
     <View style={styles.formSectionContainer}>
-      <Text style={styles.formSectionText}>Cast & Creatives</Text>
-      <View style={styles.formBlockContainer}>
-        <Text style={styles.formHeaderText}>Name</Text>
-          <View style={styles.formResponseContainer}>
-            <TextInput
-            style={styles.formResponseText}
-            placeholder="Name here"
-            placeholderTextColor={"black"}
-            value={castName}
-            textAlign={'left'}
-            onChangeText={(text) => setCastName(text)}
-            />
-          </View>
+      <View style={styles.boxContainer}>
+      <View style={styles.formResponseContainer}>
+        <TextInput
+          placeholder="Name here"
+          style={styles.formResponseText}
+          textAlign={'left'}
+          value={cast[index].name}
+          onChangeText={(text)=>fetchUsers(text) ? fetchUsers(text) : onChange(index, text, null)} />
+        <FlatList
+          style={styles.userList}
+          data={users}
+          renderItem={({ item, index }) => addUserProfile(item)}
+        />
+      </View>
       </View>
       <View style={styles.formBlockContainer}>
         <Text style={styles.formHeaderText}>Role</Text>
@@ -32,16 +36,12 @@ export default Cast = ({cast, castName, castRole}) => {
           <TextInput
           style={styles.formResponseText}
           placeholder="Role here"
-          placeholderTextColor={"black"}
-          value={castRole}
+          value={cast[index].role}
           textAlign={'left'}
-          onChangeText={(text) => setCastRole(text)}
+          onChangeText={(text) => onChange(index, null, text)}
           />
         </View>
       </View>
-      <TouchableOpacity style={styles.addButton} onPress={() => handleAdd()} >
-        <Text style={styles.buttonText}>+</Text>
-     </TouchableOpacity>
     </View>
   );
 }
