+++ screens/EditProfileScreen.tsx
@@ -1,11 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import { View, Text, TextInput, Alert, ScrollView, Keyboard ,StyleSheet, SafeAreaView, Image } from 'react-native';
-import { Ionicons } from '@expo/vector-icons';
+import { View, Text, ScrollView, Keyboard, SafeAreaView } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
+import { Ionicons } from '@expo/vector-icons';
 import * as ImagePicker from 'expo-image-picker';
 import * as ImageManipulator from 'expo-image-manipulator';
+
 import { editProfile, uploadPicture } from '../api/firebaseMethods';
+
+import styles from '../constants/Styles';
 import { useUserUpdate } from '../hooks/UserContext';
+
+import ChooseImg from '../components/ChooseImg';
+import BlockEntry from '../components/BlockEntry';
+import SingleEntry from '../components/SingleEntry';
 
 export default function EditProfileScreen({ route, navigation }) {
 
@@ -70,43 +77,11 @@ export default function EditProfileScreen({ route, navigation }) {
 
   return (
     <SafeAreaView style={styles.safeContainer}>
-      <ScrollView style={styles.container}>
+      <ScrollView style={styles.boxContainer}>
+        <ChooseImg onPress={pickImage} profPic={profile}/>
         <View style={styles.formSectionContainer}>
-          <TouchableOpacity onPress={pickImage} >
-            <Image source={profile} style={styles.imageContainer} />
-          </TouchableOpacity>
-        </View>
-        <View style={styles.formSectionContainer}>
-          <Text style={styles.formSectionText}>Display name</Text>
-          <View style={styles.formBlockContainer}>
-            <View style={styles.nameFormResponseContainer}>
-              <TextInput
-              style={styles.formResponseText}
-              placeholder=""
-              placeholderTextColor={"black"}
-              value={name}
-              textAlign={'left'}
-              onChangeText={(text) => setBio(text)}
-              />
-            </View>
-          </View>
-          <Text style={styles.formSectionText}>Edit Bio</Text>
-          <View style={styles.formBlockContainer}>
-            <View style={styles.formResponseContainer}>
-              <TextInput
-              style={styles.formResponseText}
-              placeholder="Something about you..."
-              placeholderTextColor={"black"}
-              multiline={true}
-              value={bio}
-              textAlign={'left'}
-              onChangeText={(text) => (text.split(" ").length <= 250) ?
-                (setChars(text.split(" ").length), setBio(text)) :
-                setChars(text.split(" ").length)}
-                />
-              </View>
-              <Text style={styles.formHeaderText}>{250 - chars} Words Remaining</Text>
-          </View>
+          <SingleEntry title={'Edit Name'} placeholder={''} val={name} onValUpdate={setName}/>
+          <BlockEntry title={''} placeholder={'Something about you...'} numChars={250} chars={chars} block={bio} onCharsUpdate={setChars} onBlockUpdate={setBio}/>
         </View>
         <TouchableOpacity style={styles.button} onPress={() => handlePress(id, bio)} >
           <Text style={styles.buttonText}>Save Changes</Text>
@@ -115,83 +90,3 @@ export default function EditProfileScreen({ route, navigation }) {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  safeContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    backgroundColor: 'white',
-  },
-  container: {
-    backgroundColor: 'white',
-    marginTop: 20,
-    backgroundColor: 'transparent',
-    marginHorizontal: 30,
-  },
-  imageContainer: {
-    alignSelf: 'center',
-    marginTop: 20,
-    borderRadius:60,
-    height: 120,
-    width: 120,
-  },
-  headerText: {
-    fontSize: 18,
-    paddingTop: 2,
-  },
-  formSectionContainer: {
-    marginBottom: 20,
-  },
-  formSectionText: {
-    marginTop: 10,
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  formBlockContainer: {
-    marginTop: 10,
-  },
-  formHeaderText: {
-    fontSize: 12,
-    fontWeight: '300',
-  },
-  formResponseText: {
-    fontSize: 15,
-    fontWeight: '300',
-  },
-  nameFormResponseContainer: {
-    borderRadius: 5,
-    marginVertical: 10,
-    borderBottomColor: 'gray',
-    borderBottomWidth: 0.5,
-    paddingBottom: 8,
-    paddingLeft: 20,
-    paddingRight: 20,
-  },
-  formResponseContainer: {
-    borderRadius: 5,
-    marginVertical: 10,
-    borderColor: 'gray',
-    borderWidth: 1,
-    paddingLeft: 20,
-    paddingRight: 20,
-    paddingTop: 12,
-    paddingBottom: 30,
-  },
-  button: {
-    marginTop: 20,
-    borderColor: '#c23f39',
-    borderWidth: 1,
-    borderRadius: 15,
-    paddingTop: 2,
-    paddingBottom: 2,
-    backgroundColor: 'white',
-    width: "50%",
-    alignSelf: 'center',
-  },
-  buttonText: {
-    textAlign: 'center',
-    color: '#c23f39',
-    fontSize: 16,
-  },
-
-});
