+++ screens/FavoritesScreen.tsx
@@ -1,32 +1,17 @@
 import * as React from 'react';
-import { StyleSheet } from 'react-native';
+
+import { Text, View } from '../components/Themed';
+
+import styles from '../constants/Styles';
 
 import EditScreenInfo from '../components/EditScreenInfo';
-import { Text, View } from '../components/Themed';
 
 export default function FavoritesScreen() {
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Favorites Screen</Text>
+    <View style={styles.card}>
+      <Text style={styles.titleText}>Favorites Screen</Text>
       <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
       <EditScreenInfo path="/screens/FavoritesScreen.tsx" />
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: '80%',
-  },
-});
