+++ screens/HomeScreen.tsx
@@ -1,32 +1,16 @@
 import * as React from 'react';
-import { StyleSheet } from 'react-native';
+
+import styles from '../constants/Styles';
+import { Text, View } from '../components/Themed';
 
 import Home from '../components/Home.tsx';
-import { Text, View } from '../components/Themed';
 
 export default function HomeScreen() {
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Home Screen</Text>
+    <View style={styles.card}>
+      <Text style={styles.titleText}>Home Screen</Text>
       <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
       <Home path="/screens/HomeScreen.tsx" />
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: '80%',
-  },
-});
