+++ screens/LoadingScreen.js
@@ -1,18 +1,12 @@
 import React from 'react';
-import { ActivityIndicator, StyleSheet, View } from 'react-native';
+import { ActivityIndicator, View } from 'react-native';
+
+import styles from '../constants/Styles';
 
 export default function LoadingScreen() {
   return (
-    <View style={styles.container}>
+    <View style={styles.stdContainer}>
       <ActivityIndicator size='large' />
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
