+++ screens/SearchScreen.tsx
@@ -1,10 +1,15 @@
 import React, { useEffect, useState } from 'react';
-import { StyleSheet, TextInput, FlatList, TouchableOpacity } from 'react-native';
+import { TouchableOpacity } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
 import * as firebase from 'firebase';
 import { follow, unfollow } from '../api/firebaseMethods'
-import { Ionicons } from '@expo/vector-icons';
+
+import styles from '../constants/Styles';
 import { Text, View } from '../components/Themed';
 import { useUser, useUserUpdate } from '../hooks/UserContext';
+
+import SearchBar from '../components/SearchBar';
 
 export default function MessagesScreen({ navigation }) {
   const [users, setUsers] = useState([]);
@@ -51,59 +56,10 @@ export default function MessagesScreen({ navigation }) {
   }
 
   return (
-    <View style={styles.container}>
-      <View style={styles.searchBarContainer}>
-        <TextInput
-          placeholder="Search!"
-          style={styles.searchBar}
-          textAlign={'center'}
-          onChangeText={(search)=>fetchUsers(search)} />
-        <Text style={styles.resultsHeaderText}>Results:</Text>
-        <FlatList
-          style={styles.userList}
-          data={users}
-          renderItem={({ item, index }) => getUserProfile(item)} />
-      </View>
-    </View>
+    <SearchBar
+      users={users}
+      onTextUpdate={fetchUsers}
+      getUserProfile={getUserProfile}
+    />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-
-  },
-  searchBarContainer: {
-    width: '70%',
-    marginTop: 40,
-  },
-  resultsHeaderText: {
-    fontWeight: '400',
-    fontSize: 18,
-    marginTop: 40,
-  },
-  searchBar: {
-    borderBottomColor: 'black',
-    borderBottomWidth: 1,
-    fontSize: 18,
-  },
-  userList: {
-    marginTop: 20,
-  },
-  userItemContainer: {
-    borderRadius: 8,
-    paddingTop: 8,
-    paddingBottom: 8,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  userName: {
-    marginVertical: 1,
-    fontSize: 18,
-  },
-  userHandle: {
-    color: 'gray',
-  }
-
-});
