+++ screens/WelcomeScreen.js
@@ -1,5 +1,5 @@
-import {ImageBackground, StyleSheet, View, Text} from 'react-native';
 import React from 'react';
+import {ImageBackground, StyleSheet, View, Text} from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 
 export default function WelcomeScreen ({navigation}) {
